requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
urllib3==1.26.18
coverage==7.3.1
pyyaml==6.0